+++ src/signposting/__init__.py
@@ -1,6 +1,6 @@
 """Finding signposting in FAIR resources.
 
-This library helps client to discover links that follow the 
+This library helps client to discover links that follow the
 `FAIR`_ `signposting`_ conventions.
 
 This can then be used to navigate between:
@@ -12,15 +12,15 @@ This can then be used to navigate between:
 
 The library works by inspecting the HTTP messages for
 ``Link:`` headers from a given URI with :meth:`find_signposting_http`, which
-which categorize them by their ``rel`` `Link relation`_ into a 
+which categorize them by their ``rel`` `Link relation`_ into a
 :class:`Signposting` object with absolute URIs.
 
 It is up to the clients of this library to decide how to further
-navigate or retrieve the associated resources, e.g. using a 
+navigate or retrieve the associated resources, e.g. using a
 RDF library like :mod:`rdflib`.
 
 Future versions of this library may also provide ways to discover
-FAIR signposting in HTML ``<link>`` annotations and in 
+FAIR signposting in HTML ``<link>`` annotations and in
 `linkset`_ documents.
 
 .. _signposting: https://signposting.org/conventions/