+++ src/signposting/linkheader.py
@@ -61,7 +61,7 @@ def _optional_link(parsedLinks: ParsedLinks, rel: str) -> Optional[Link]:
 
 def _link_attr(link: Link, key: str) -> Optional[str]:
     """Look up an optional link attribute with given `key` from a `Link`.
-    
+
     Return the attribute value, or ``None`` if the link attribute was not found.
 
     This is a workaround as `Link` exposes ``__getitem__`` and ``__contains__`` but not ``Dict.get()``
@@ -71,27 +71,27 @@ def _link_attr(link: Link, key: str) -> Optional[str]:
     return None
 
 def linkToSignpost(link: Link, rel: LinkRel, context_url: str = None) -> Signpost:
-    return Signpost(rel, link.target, 
+    return Signpost(rel, link.target,
         _link_attr(link, "type"),
         _link_attr(link, "profile"),
         context_url, link)
 
 def linksToSignposting(links: List[Link], context_url: str = None) -> Signposting:
-        """Initialize Signposting object for a given `ParsedLinks` 
+        """Initialize Signposting object for a given `ParsedLinks`
         as discovered from the (optional) `context_url` base URL.
         """
         signposts: List[Signpost] = []
         for l in links:
             # TODO: Check if context_url matches "anchor"
             for rel in l.rel:
-                if rel in SIGNPOSTING:                    
+                if rel in SIGNPOSTING:
                     signposts.append(linkToSignpost(l, LinkRel(rel), context_url))
         return Signposting(context_url, signposts)
 
 def _absolute_attribute(k: str, v: str, baseurl: str) -> Tuple[str, str]:
     """Ensure link attribute value uses absolute URI, resolving from the baseurl.
 
-    Currently this will mean the `context`_ attribute ``anchor`` 
+    Currently this will mean the `context`_ attribute ``anchor``
     and ``profile`` will be rewritten.
 
     .. _context: https://datatracker.ietf.org/doc/html/rfc8288#section-3.2
@@ -102,7 +102,7 @@ def _absolute_attribute(k: str, v: str, baseurl: str) -> Tuple[str, str]:
 
 
 def find_signposting(headers: List[str], baseurl: str = None) -> Signposting:
-    """Find signposting among HTTP Link headers. 
+    """Find signposting among HTTP Link headers.
 
     Optionally make the links absolute according to the base URL.